+++ src/app/core/model/json_map.ts
