+++ src/app/features/todos/domain/infrastructure/todos_api.ts
@@ -83,13 +83,3 @@ export abstract class TodosApi {
    */
   abstract close(): Promise<void>;
 }
-
-/**
- * Custom exception thrown when a todo is not found.
- */
-export class TodoNotFoundException extends Error {
-  constructor() {
-    super('Todo not found');
-    this.name = 'TodoNotFoundException';
-  }
-}
